[task] Add middleware and validators

+++ src/products/products.controller.ts
@@ -1,6 +1,8 @@
-import { Controller, Param, Body, Get, Delete, Post } from '@nestjs/common';
+import { Controller, Param, Body, Get, Delete, Post, Put, NotFoundException } from '@nestjs/common';
 import { ProductsService } from './products.service';
+import { ParseUUIDPipe } from '@nestjs/common';
 import { CreateProductDTO } from './dtos/create-product.dto';
+import { UpdateProductDTO } from './dtos/update-product.dto';
 
 @Controller('products')
 export class ProductsController {
@@ -13,12 +15,16 @@ export class ProductsController {
     }
 
     @Get('/:id')
-    getById(@Param('id') id: string): any {
-        return this.productsService.getById(id);
+    getById(@Param('id', new ParseUUIDPipe()) id: string): any {
+        const prod = this.productsService.getById(id);
+        if (!prod) throw new NotFoundException('Product not found')     // error handling https://docs.nestjs.com/exception-filters#built-in-http-exceptions
+        return prod;
     }
 
     @Delete('/:id')
-    deleteById(@Param('id') id: string) {
+    deleteById(@Param('id', new ParseUUIDPipe()) id: string) {
+        if (!this.productsService.getById(id))
+            throw new NotFoundException('Product not found')
         this.productsService.deleteById(id);
         return { success: true };
     }
@@ -28,6 +34,18 @@ export class ProductsController {
         return this.productsService.create(productData);
     }
 
+    @Put('/:id')
+    update(
+        @Param('id', new ParseUUIDPipe()) id: string,
+        @Body() productData: UpdateProductDTO,
+    ) {
+        if (!this.productsService.getById(id))
+            throw new NotFoundException('Product not found')
+
+        this.productsService.updateById(id, productData);
+        return { success: true };
+    }
+
 }
 
 /*  DTO (Data Object Model) - 
@@ -35,4 +53,6 @@ export class ProductsController {
         NestJS
         rekomenduje, aby tworzyć specjalne oddzielne struktury opisujące dane otrzymywane w
         przypadku requestu do dodawania elementu, usuwania itd
+
+        pipes - https://docs.nestjs.com/pipes
 */
